Auth endpoints

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Task } from './task.entity'
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto'
+import { User } from '../auth/user.entity'
 
 @Injectable()
 export class TasksService {
@@ -13,12 +14,14 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filterDto)
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filterDto, user)
   }
 
-  async getTaskById(id: number): Promise<Task | undefined> {
-    const task = await this.taskRepository.findOne(id)
+  async getTaskById(id: number, user: User): Promise<Task | undefined> {
+    const task = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    })
 
     if (!task) {
       throw new NotFoundException(`Task with ID ${id} not found`)
@@ -27,16 +30,20 @@ export class TasksService {
     return task
   }
 
-  async deleteTaskById(id: number) {
-    const result = await this.taskRepository.delete(id)
+  async deleteTaskById(id: number, user: User) {
+    const result = await this.taskRepository.delete({ id, userId: user.id })
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID ${id} not found`)
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task | void> {
-    const task = await this.getTaskById(id)
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task | void> {
+    const task = await this.getTaskById(id, user)
 
     if (task) {
       task.status = status
@@ -46,7 +53,7 @@ export class TasksService {
     }
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user)
   }
 }
